frame/mmr: completely move primitives out of frame
