+++ client/beefy/src/lib.rs
@@ -27,10 +27,10 @@ use sp_api::ProvideRuntimeApi;
 use sp_blockchain::HeaderBackend;
 use sp_consensus::SyncOracle;
 use sp_keystore::SyncCryptoStorePtr;
+use sp_mmr_primitives::MmrApi;
 use sp_runtime::traits::Block;
 
 use beefy_primitives::{BeefyApi, MmrRootHash};
-use sp_mmr_primitives::MmrApi;
 
 use crate::notification::{BeefyBestBlockSender, BeefySignedCommitmentSender};
