--- frame/merkle-mountain-range/primitives/Cargo.toml
@@ -1,44 +1,0 @@
-[package]
-name = "pallet-mmr-primitives"
-version = "4.0.0-dev"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2021"
-license = "Apache-2.0"
-homepage = "https://substrate.io"
-repository = "https://github.com/paritytech/substrate/"
-description = "FRAME Merkle Mountain Range primitives."
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false }
-log = { version = "0.4.14", default-features = false }
-serde = { version = "1.0.136", optional = true, features = ["derive"] }
-
-sp-api = { version = "4.0.0-dev", default-features = false, path = "../../../primitives/api" }
-sp-core = { version = "6.0.0", default-features = false, path = "../../../primitives/core" }
-sp-debug-derive = { version = "4.0.0", default-features = false, path = "../../../primitives/debug-derive" }
-sp-mmr-primitives = { version = "4.0.0-dev", default-features = false, path = "../../../primitives/merkle-mountain-range" }
-sp-runtime = { version = "6.0.0", default-features = false, path = "../../../primitives/runtime" }
-sp-std = { version = "4.0.0", default-features = false, path = "../../../primitives/std" }
-
-frame-system = { version = "4.0.0-dev", default-features = false, path = "../../system" }
-
-[dev-dependencies]
-hex-literal = "0.3"
-
-[features]
-default = ["std"]
-std = [
-	"codec/std",
-	"log/std",
-	"serde",
-	"sp-api/std",
-	"sp-core/std",
-	"sp-debug-derive/std",
-	"sp-mmr-primitives/std",
-	"sp-runtime/std",
-	"sp-std/std",
-	"frame-system/std",
-]
