+++ frame/merkle-mountain-range/rpc/Cargo.toml
@@ -25,7 +25,5 @@ sp-core = { version = "6.0.0", path = "../../../primitives/core" }
 sp-mmr-primitives = { version = "4.0.0-dev", default-features = false, path = "../../../primitives/merkle-mountain-range" }
 sp-runtime = { version = "6.0.0", path = "../../../primitives/runtime" }
 
-pallet-mmr-primitives = { version = "4.0.0-dev", path = "../primitives" }
-
 [dev-dependencies]
 serde_json = "1.0.79"
