+++ frame/merkle-mountain-range/src/mmr/mod.rs
@@ -19,7 +19,7 @@ mod mmr;
 pub mod storage;
 pub mod utils;
 
-use sp_mmr_primitives::FullLeaf;
+use sp_mmr_primitives::{DataOrHash, FullLeaf};
 use sp_runtime::traits;
 
 pub use self::mmr::{verify_leaf_proof, Mmr};
@@ -28,7 +28,7 @@ pub use self::mmr::{verify_leaf_proof, Mmr};
 pub type NodeOf<T, I, L> = Node<<T as crate::Config<I>>::Hashing, L>;
 
 /// A node stored in the MMR.
-pub type Node<H, L> = pallet_mmr_primitives::DataOrHash<H, L>;
+pub type Node<H, L> = DataOrHash<H, L>;
 
 /// Default Merging & Hashing behavior for MMR.
 pub struct Hasher<H, L>(sp_std::marker::PhantomData<(H, L)>);
