+++ frame/merkle-mountain-range/src/tests.rs
@@ -19,11 +19,11 @@ use crate::{mmr::utils, mock::*, *};
 
 use frame_support::traits::OnInitialize;
 use mmr_lib::helper;
-use pallet_mmr_primitives::{Compact, Proof};
 use sp_core::{
 	offchain::{testing::TestOffchainExt, OffchainDbExt, OffchainWorkerExt},
 	H256,
 };
+use sp_mmr_primitives::{Compact, Proof};
 
 pub(crate) fn new_test_ext() -> sp_io::TestExternalities {
 	frame_system::GenesisConfig::default().build_storage::<Test>().unwrap().into()
