+++ primitives/merkle-mountain-range/Cargo.toml
@@ -19,6 +19,7 @@ serde = { version = "1.0.136", optional = true, features = ["derive"] }
 sp-api = { version = "4.0.0-dev", default-features = false, path = "../api" }
 sp-core = { version = "6.0.0", default-features = false, path = "../core" }
 sp-debug-derive = { version = "4.0.0", default-features = false, path = "../debug-derive" }
+sp-runtime = { version = "6.0.0", default-features = false, path = "../runtime" }
 sp-std = { version = "4.0.0", default-features = false, path = "../std" }
 
 [dev-dependencies]
@@ -33,5 +34,6 @@ std = [
 	"sp-api/std",
 	"sp-core/std",
 	"sp-debug-derive/std",
+	"sp-runtime/std",
 	"sp-std/std",
 ]
